requirements.txt
feedparser==6.0.2  
PyGithub==1.51